requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via pytest
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via -r requirements.in
google-api-core==2.10.2
    # via google-api-python-client
google-api-python-client==2.66.0
    # via -r requirements.in
google-auth==2.14.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==0.7.1
    # via -r requirements.in
googleapis-common-protos==1.57.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
numpy==1.23.5
    # via
    #   -r requirements.in
    #   pandas
    #   scipy
oauthlib==3.2.2
    # via requests-oauthlib
oura==1.2.4
    # via -r requirements.in
packaging==21.3
    # via pytest
pandas==1.5.2
    # via oura
pluggy==1.0.0
    # via pytest
protobuf==4.21.9
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.2
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
pytest==7.2.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2022.6
    # via pandas
requests==2.28.1
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   oura
rsa==4.9
    # via google-auth
scipy==1.9.3
    # via -r requirements.in
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
typing-extensions==4.4.0
    # via pydantic
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.13
    # via requests
